\name{get.signs.M}
\alias{get.signs.M}
\title{
Vetorizes connection sign matrix.
}
\description{
Stores a matrix of connection signs to a vector.
}
\usage{get.signs.M(MAT)}
\arguments{
\item{MAT}{matrix of connection signs that contains -1, 1 or 0}

}
\value{
\item{vec.out}{vectorized MAT matrix}
}
\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}

