\name{mat.to.laplacian}
\alias{mat.to.laplacian}
\title{
Computes laplacian given some matrix M. 
}
\description{
Computes laplacian given some matrix M.
}
\usage{mat.to.laplacian(M1,type)}
\arguments{
\item{M1}{\eqn{p \times p} matrix}
\item{type}{Laplacian types. "normalized" (default) gives normalized laplacian, "combinatorial" gives combinatorial.}
}
\value{
\item{L}{Laplacian}
}
\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}

