\name{get.signs.M}
\alias{get.signs.M}
\title{
A function that vetorizes connetion sign matrix.
}
\description{
Stores a matrix of connection signs into a vector.
}
\usage{get.signs.M(MAT)}
\arguments{
\item{MAT}{A matrix of connection signs that contains -1, 1 or 0}

}
\value{
\item{vec.out}{Vectorized MAT matrix}
}
\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}

