\name{LassoNet-package}
\alias{LassoNet-package}
\docType{package}
\title{
LassoNet: package for 3CoSE algorithm. 

}
\description{
LassoNet contains functions to estimate a penalized regression model using 3CoSE algorithm described in the paper Weber, Striaukas, Schumacher and Binder (2018). The main function of the package is the function lasso.net.grid, see the example below.}
\details{
\tabular{ll}{
Package: \tab LassoNet\cr
Type: \tab Package\cr
Version: \tab 0.8.3\cr
Date: \tab 2019-12-16 \cr
License: \tab Open source \cr
}
}

\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}
\references{Weber, M., Striaukas, J., Schumacher, M., Binder, H. "Network-Constrained Covariate Coefficient and Connection Sign Estimation" (2018) <doi:10.2139/ssrn.3211163>}

\seealso{
\code{\link[Rcpp:Rcpp-package]{Rcpp}}, \code{\link[glmnet:glmnet-package]{glmnet}}
}
