\name{LassoNet-package}
\alias{LassoNet-package}
\docType{package}
\title{
LassoNet: package for 3CoSE algorithm. 

}
\description{
LassoNet contains functions to estimate a penalized regression model using 3CoSE algorithm described in the paper Weber, Striaukas, Schumacher and Binder (2018). The main function of the package is the function lasso.net.grid, see the example below.}
\details{
\tabular{ll}{
Package: \tab LassoNet\cr
Type: \tab Package\cr
Version: \tab 0.8.3\cr
Date: \tab 2019-12-16 \cr
License: \tab Open source \cr
}
}

\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}
\references{
Weber, M., Schumacher, M., & Binder, H. (2014) \emph{Regularized Regression Incorporating Network Information: Simultaneous Estimation of Covariate Coefficients and Connection Signs}  

Weber, M., Striaukas, J., Schumacher, M., & Binder, H. (2018) \emph{Network Constrained Covariate Coefficient and Connection Sign Estimation}
}
\seealso{
\code{\link[Rcpp:Rcpp-package]{Rcpp}}, \code{\link[glmnet:glmnet-package]{glmnet}}
}
