\name{mat.to.laplacian}
\alias{mat.to.laplacian}
\title{
Computes Laplacian matrix. 
}
\description{
Computes Laplacian matrix. 
}
\usage{mat.to.laplacian(M1,type)}
\arguments{
\item{M1}{\eqn{p \times p} matrix}
\item{type}{Laplacian types: 1) "normalized" (default) - normalized Laplacian, 2) "combinatorial" - combinatorial Laplacian}
}
\value{
\item{L}{Laplacian}
}
\author{
Maintainer: Jonas Striaukas <jonas.striaukas@gmail.com>
}

